fix /admin/regenerate (with selectivity too)
The problem: regenerate_all() would ignore the desire to regenerate (as
it was conditioned on the BLOGGART_VERSION).

Any fix for that would be compounded by the fact that post regeneration
would be executed again by try_post_deploy(). Nor can we skip it and go
straight to post_deploy() due to the VersionInfo entity argument; see
a8261bd (don't pass a tuple to post_deploy()).

So, to have our cake and eat it too, we let kwargs trickle through the
post-depoly tasks from try_post_deploy(), and force regenerate_all() to
call the PostRegenerator.